notifications: refactor email code for easier import
We only want to import the key functions and hide the complexity.
